fix: balance album list metadata spacing
